requirements.txt
absl-py==0.1.11
astor==0.6.2
audioread==2.1.5
backcall==0.1.0
bleach==1.5.0
cffi==1.11.5
cloudpickle==0.5.3
cycler==0.10.0
dask==0.17.5
decorator==4.2.1
Django==2.0.1
entrypoints==0.2.3
enum34==1.1.6
gast==0.2.0
grpcio==1.10.0
h5py==2.7.0
html5lib==0.9999999
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
jedi==0.11.1
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.0.5
kiwisolver==1.0.1
librosa==0.5.1
llvmlite==0.22.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.4.1
numba==0.37.0
numpy==1.14.2
opencv-contrib-python==3.4.1.15
opencv-python==3.2.0.7
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-speech-features==0.5
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
resampy==0.2.0
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.7.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
SoundFile==0.9.0.post1
tensorboard==1.7.0
tensorflow==1.7.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tflearn==0.3.2
Theano==1.0.1
toolz==0.9.0
tornado==5.0.2
tqdm==4.19.7
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.0
django-watson==1.5.2
python-env==1.0.0
psycopg2==2.7.5
gensim
elasticsearch-dsl>=2.1,<3.0
redis
djangorestframework
gunicorn